Share android/ohos HostTrait
